+++ words/models/__init__.py
@@ -1,0 +1,3 @@
+from .word_models import *
+from .article_models import *
+from .article_word_models import *
