+++ words/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from words.models import Word, Article, ArticleWord
+from words.models import *
 
 class WordSerializer(serializers.ModelSerializer):
   class Meta:
