+++ words/views/__init__.py
@@ -1,0 +1,2 @@
+from .word_views import *
+from .article_views import *
